chore(deps): update rand requirement from 0.8.5 to 0.9.2
- Update rand from 0.8.5 to 0.9.2 in dev-dependencies
- Fix rand::random::<usize>() compatibility issue in api_polling_scraping.rs
- rand 0.9 removed StandardUniform support for usize
- Changed to use u32 and cast to usize
- All tests passing

+++ examples/api_polling_scraping.rs
@@ -111,7 +111,7 @@ impl DataStore {
             city: city.to_string(),
             temperature: 15.0 + (rand::random::<f64>() * 20.0),
             conditions: vec!["Sunny", "Cloudy", "Rainy", "Partly Cloudy"]
-                [rand::random::<usize>() % 4]
+                [(rand::random::<u32>() % 4) as usize]
                 .to_string(),
             timestamp: Utc::now(),
         };
